fix(Logger): move aut token to the beginning of the log message
Each log message contains the authentication token, and logentries
allows the token to be everywhere in the message sent.
Previously the token was at the end of the message.
In case the message contained another UUID, logentries used it as
an authentication token and skipped the real one that was at the end of
the message.
For this reason, all the logs with a UUID inside where not saved in
logentries.